Fixing delete in progress
Able to delete a user in the UserDB if they're authenticated.

Needs refractoring because the user should be deleted in the login & token table as well

+++ UserApplication/UserCrud.cs
@@ -2,6 +2,9 @@
 using Domain;
 using Domain.DTOs;
 using Microsoft.AspNetCore.Identity;
+using Newtonsoft.Json;
+using System.Net.Http;
+using System.Security.Cryptography.X509Certificates;
 using UserApplication.Interfaces;
 using UserInfrastructure.Interfaces;
 
@@ -11,11 +14,16 @@ public class UserCrud : IUserCrud
 {
     private readonly IUserRepo _userRepo;
     private readonly IMapper _mapper;
+    private readonly IHttpClientFactory _httpClientFactory;
+    private readonly string _authServiceUrl;
 
-    public UserCrud(IUserRepo userRepo, IMapper mapper)
+
+    public UserCrud(IUserRepo userRepo, IMapper mapper, IHttpClientFactory httpClientFactory)
     {
         _userRepo = userRepo;
         _mapper = mapper;
+        _authServiceUrl = "https://localhost:7227/Auth";
+        _httpClientFactory = httpClientFactory;
     }
 
     public async Task<User> AddUserAsync(UserDto userDto)
@@ -24,9 +32,15 @@ public class UserCrud : IUserCrud
         return await _userRepo.CreateUserAsync(user);
     }
 
-    public async Task DeleteUserAsync(int userId)
+    public async Task<bool> DeleteUserAsync(int userId, int requesterUserId)
     {
+        if (userId != requesterUserId || !await CheckUserHasActiveTokenAsync(requesterUserId))
+        {
+            return false;
+        }
+
         await _userRepo.DeleteUserAsync(userId);
+        return true;
     }
 
     public async Task<User> GetUserAsync(int userId)
@@ -60,5 +74,22 @@ public class UserCrud : IUserCrud
     public async Task<User> GetUserByEmail(string email)
     {
         return await _userRepo.GetUserByEmailAsync(email);
+    }
+
+    private async Task<bool> CheckUserHasActiveTokenAsync(int userId)
+    {
+        var httpClient = _httpClientFactory.CreateClient();
+        var authServiceUrl = _authServiceUrl; 
+
+        var response = await httpClient.GetAsync($"{authServiceUrl}/{userId}/hasActiveToken");
+
+        if (!response.IsSuccessStatusCode)
+        {
+            return false;
+        }
+
+        var responseContent = await response.Content.ReadAsStringAsync();
+        var tokenStatus = JsonConvert.DeserializeObject<TokenStatusDto>(responseContent);
+        return tokenStatus?.IsActive ?? false;
     }
 }
